pyproject.toml
[project]
name = "chronolog-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.12"
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.1",
]
